Add reservation fix
