+++ src/index.html
@@ -43,10 +43,9 @@
                 <button id="add-comment">Comment</button>
             </div>
             <div class="section" id="reservation-section">
-                <h3>
-                    Reservation <span><i id="close-btn" class="fa fa-times"></i></span>
+                <h3 id="res-counter">
+                    
                 </h3>
-                Stats
                 <ul class="reservation-list">
     
                 </ul>
