+++ src/modules/api.js
@@ -1,4 +1,4 @@
-const invAPI = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/dSv5DdTGl6SZHdXDAlEr/';
+const invAPI = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/wr5n1QxD7pfK5gphgr0z/';
 const pokeAPI = 'https://pokeapi.co/api/v2/pokemon/';
 
 const pokeArr = [];
