+++ src/modules/reservation.js
@@ -25,43 +25,21 @@ const addReservation = async (id, name, dateStart, dateEnd) => {
 
 /* Display functions */
 
-const newReservation = document.querySelector('#add-reservation');
-const reservationForm = document.querySelector('.reservation-form');
-const reserveBtn = document.querySelector('.reserve-button');
-const reservationPopup = document.querySelector('#reservation-popup');
-const closeBtn = document.querySelector('#close-btn');
-const list = document.querySelector('#reservation-list');
+const resList = document.querySelector('.reservation-list');
+const totalReservations = document.querySelector("#res-counter");
 
-const displayReservation = async () => {
-  const reservationList = await getReservation('item2');
-  reservationList.forEach((item) => {
-    list.innerHTML += `<li class="reservationName"> ${item.username}: ${item.date_start} ${item.date_end}`;
-  });
+const displayReservation = async (id) => {
+  const reservationData = await getReservation(id);
+  resList.innerHTML = '';
+  if (reservationData.error) {
+    totalReservations.textContent = `Reservations (0):`
+  } else {
+    totalReservations.textContent = `Reservations (${reservationData.length}):`
+    reservationData.forEach((item) => {
+      resList.innerHTML += `<li class="reservation"> ${item.username}: ${item.date_start} ${item.date_end}`;
+    });
+  }
 };
 
-reserveBtn.addEventListener('click', () => {
-  reservationPopup.classList.remove('reservation-overlay');
-});
 
-closeBtn.addEventListener('click', () => {
-  reservationPopup.classList.add('reservation-overlay');
-});
-
-newReservation.addEventListener('click', (e) => {
-  e.preventDefault;
-  const reserverName = document.querySelector('#reservationName').value;
-  const startDate = document.querySelector('#startDate').value;
-  const endDate = document.querySelector('#endDate').value;
-  getReservation('item2');
-  if (!reserverName || !startDate || !endDate) {
-    e.preventDefault();
-    console.log('oops!');
-  } else {
-    addReservation('item2', reserverName, startDate, endDate);
-    //reservation-popup.innerHTML = '';
-    reservationForm.reset();
-  }
-});
-
-
-export { getReservation, addReservation };
+export { getReservation, addReservation , displayReservation };
